added auto scrolling, time bar, and all day tasks
