+++ todo-app/Views/TimeIndicatorViews.swift
@@ -1,0 +1,4 @@
+// This file is no longer needed as the components have been moved to individual files:
+// - TimeIndicatorView.swift
+// - FixedTimeIndicator.swift
+// - TimeIndicatorPositioner.swift
